+++ library/src/main/jni/image/image.c
@@ -63,7 +63,7 @@ static int8_t get_format(Stream* stream, uint8_t* magic) {
 
 void decode(Stream* stream, bool partially, bool* animated, void** image) {
   uint8_t magic[IMAGE_MAGIC_NUMBER_BYTE_COUNT];
-  int8_t format;
+  int32_t format;
 
   // Get image format
   if (!stream->mark(stream, IMAGE_MAGIC_NUMBER_BYTE_COUNT)) {
@@ -137,7 +137,7 @@ bool decode_info(Stream* stream, ImageInfo* info) {
 bool decode_buffer(Stream* stream, bool clip, uint32_t x, uint32_t y, uint32_t width,
     uint32_t height, int32_t config, uint32_t ratio, BufferContainer* container) {
   uint8_t magic[2];
-  int8_t format;
+  int32_t format;
 
   // Get image format
   if (!stream->mark(stream, IMAGE_MAGIC_NUMBER_BYTE_COUNT)) {
