+++ library/src/main/jni/image/image_info.h
@@ -13,7 +13,7 @@
 typedef struct {
   uint32_t width;
   uint32_t height;
-  int8_t format;
+  int32_t format;
   bool opaque;
   int32_t frame_count;
 } ImageInfo;
